requirements.txt
gym==0.12.5
imageio>=2.3.0
scikit-image>=0.14.0
tensorboardX==2.0
numpy==1.14.0
setproctitle==1.1.10
tqdm==4.45.0
Pytorch >= 1.2


